requirements.txt
python-dotenv
psutil
pymongo
requests
PyNaCl
bs4
html5lib
discord.py==2.5.2
wavelink>=3.4.0
humanfriendly>=10.0
gtts>=2.3.2
cashfree-pg>=4.0.0
google-generativeai
pytz
dateparser
google-generativeai